+++ Addon/!mMT_MediaPack/media/core.lua
@@ -657,6 +657,7 @@ mAddStatusbar("MaUIv3 LEFT", "MaUIv3Left.tga")
 mAddStatusbar("MaUIv3 RIGHT", "MaUIv3Right.tga")
 mAddStatusbar("mMT Blank", "mMT_Blank.tga")
 mAddStatusbar("mMT Target", "mMT_Target.tga")
+mAddStatusbar("mMT Dark", "mMT_Dark.tga")
 
 -- Backgrounds
 mAddBackground("mMediaTag BG1", "bg1.tga")