+++ Addon/ElvUI_mMediaTag/core/changelog/new.lua
@@ -1,5 +1,5 @@
-mMT.Changelog[356]  = {
-	DATE = "there is nothing new here",
+mMT.Changelog[111]  = {
+	DATE = "Dev not",
 	IMPORTANT = { "there is nothing new here" },
 	NEW = { "there is nothing new here" },
 	UPDATE = { "there is nothing new here" },