+++ Addon/ElvUI_mMediaTag/localization/koKR.lua
@@ -21,7 +21,7 @@ L["Adds an Extra Icon to the Castbar."] = true
 L["After setting the position, the interface must be reloaded once with /rl."] = true
 L["All Dungeon Teleports"] = true
 L["Alliance: "] = true
-L["Allways withe"] = true
+L["Always withe"] = true
 L["Alpha"] = true
 L["Anchor Point"] = true
 L["Anchor point for the extra Icon."] = true
@@ -807,6 +807,18 @@ L["|cffFFFFFFControl + Left Click:|r Change Loadout"] = true
 L["|cffFFFFFFLeft Click:|r Change Talent Specialization"] = true
 L["|cffFFFFFFRight Click:|r Change Loot Specialization"] = true
 L["|cffFFFFFFShift + Left Click:|r Show Talent Specialization UI"] = true
+
+L["PROFESSION_ALCHEMY"] = "연금술";
+L["PROFESSION_BLACKSMITHING"] = "대장기술";
+L["PROFESSION_ENCHANTING"] = "마법부여";
+L["PROFESSION_ENGINEERING"] = "기계공학";
+L["PROFESSION_HERBALISM"] = "약초채집";
+L["PROFESSION_INSCRIPTION"] = "주문각인";
+L["PROFESSION_JEWELCRAFTING"] = "보석세공";
+L["PROFESSION_LEATHERWORKING"] = "가죽세공";
+L["PROFESSION_MINING"] = "채광";
+L["PROFESSION_SKINNING"] = "무두질";
+L["PROFESSION_TAILORING"] = "재봉술";
 
 -- new
 L["Reaction"] = true