+++ Addon/ElvUI_mMediaTag/localization/ptBR.lua
@@ -21,7 +21,7 @@ L["Adds an Extra Icon to the Castbar."] = true
 L["After setting the position, the interface must be reloaded once with /rl."] = true
 L["All Dungeon Teleports"] = true
 L["Alliance: "] = true
-L["Allways withe"] = true
+L["Always withe"] = true
 L["Alpha"] = true
 L["Anchor Point"] = true
 L["Anchor point for the extra Icon."] = true
@@ -807,6 +807,18 @@ L["|cffFFFFFFControl + Left Click:|r Change Loadout"] = true
 L["|cffFFFFFFLeft Click:|r Change Talent Specialization"] = true
 L["|cffFFFFFFRight Click:|r Change Loot Specialization"] = true
 L["|cffFFFFFFShift + Left Click:|r Show Talent Specialization UI"] = true
+
+L["PROFESSION_ALCHEMY"] = "Alquimia";
+L["PROFESSION_BLACKSMITHING"] = "Ferraria";
+L["PROFESSION_ENCHANTING"] = "Encantamento";
+L["PROFESSION_ENGINEERING"] = "Engenharia";
+L["PROFESSION_HERBALISM"] = "Herborismo";
+L["PROFESSION_INSCRIPTION"] = "Escrivania";
+L["PROFESSION_JEWELCRAFTING"] = "Joalheria";
+L["PROFESSION_LEATHERWORKING"] = "Couraria";
+L["PROFESSION_MINING"] = "Mineração";
+L["PROFESSION_SKINNING"] = "Esfolamento";
+L["PROFESSION_TAILORING"] = "Alfaiataria";
 
 -- new
 L["Reaction"] = true